container x86-64: always enable debug tweaks
Previously we set EXTRA_IMAGE_FEATURES = debug-tweaks in local.conf
depending on whether we build a development image. The previous line in
container_qemux86-64.inc enabled debug-tweaks if not set or did nothing
if debug-tweaks was set anyways. Thus debug-tweaks was always enabled.
We should discuss if this is a good idea, however if disabled, it breaks
the intergration tests as we cannot access the container via SSH.

This commmit also always set debug-tweaks in the container.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>