conf/multiconfig/installer: increase initramfs
Since, latest changes to cml-base.conf in meta-gyroidos the initramfs
for non DEVELOPMENT_BUILD targets is set to 35000K which is to small for
the installer image. Since the installer needs also debug-tweaks, we
set the DEVELOPMENT_BUILD flag to 'y'. This uses the grater image size.

Fixes following build error:

  ERROR: mc:installer:gyroidos-installer-initramfs-1.0-r0 do_image_cpio: \
  The initramfs size 79810(K) exceeds INITRAMFS_MAXSIZE: 35000(K)

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>